komi-republic-django/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2

# PostgreSQL support (optional, for production)
# Uncomment if you need PostgreSQL:
# For Python 3.11-3.13:
# psycopg2-binary==2.9.10
# For Python 3.14+:
# psycopg[binary]==3.2.3
